+++ .github/release-notes.md
@@ -1,4 +1,5 @@
 ## Changes:
 
+- **BREAKING CHANGE:** The `-n`, `--player-name` argument has been removed. Use `-a`, `--allowlist-add` instead.
 - Added support for configuring the program via a configuration file.
 - Print debug logs with `-b`, `--debug-log`.